% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yearly_rainfall.R
\name{aggregate_rainfall}
\alias{aggregate_rainfall}
\title{Get the summed yearly rainfall}
\usage{
aggregate_rainfall(
  path_ncdf = "/mnt/CEPH_PROJECTS/Proslide/PREC_GRIDS_updated/",
  polygon = NULL,
  years = 1980:2018,
  by = "year",
  daily_fun = NULL,
  monthly_fun = NULL,
  return.list = FALSE,
  aggre_fun = "mean"
)
}
\arguments{
\item{return.list}{if \code{True} it will return a list of stars oject. If \code{false} will return one stars object (if \code{by} = \code{year}),
and a list of stars-objects if \code{by = 'month'}}

\item{aggre_fun}{If you pass a polygon, the values need to be aggregated somehow}

\item{fun}{How to combine the individual values. E.G. the mean for each pixel for the years 1980, 1981 and 1982 requires the
the fun to be "mean". The maximum value for each month in the years 80,81,82 requires the \code{by}-parameter to be \code{month}
and the \code{fun}-parameter to be \code{max}}
}
\description{
This function takes either a polygon, or nothing (than it's pixel based) and calculates the sum of rainfall for each polygon
(each pixel) for each year
}
