% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{bozen}
\alias{bozen}
\title{Spatial object of the city of bozen}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 1 columns.
}
\usage{
bozen
}
\description{
Spatial object of the city of bozen
}
\keyword{datasets}
