% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_of_slides_in_poly.R
\name{dates_of_slides_in_poly}
\alias{dates_of_slides_in_poly}
\title{For each polygon with slide it keeps one row per date}
\usage{
dates_of_slides_in_poly(joined_points_on_poly)
}
\arguments{
\item{joined_points_on_poly}{An object of sf where the points were joined on the polygon}
}
\description{
For each polygon with slide it keeps one row per date
}
