% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_rainfall.R
\name{ex_rainfall}
\alias{ex_rainfall}
\title{Extract rainfall data}
\usage{
ex_rainfall(
  data_path = "\\\\\\\\projectdata.eurac.edu/projects/Proslide/PREC_GRIDS_updated/",
  spatial.obj = NULL,
  fun = NULL,
  date = NULL,
  nc_var = "precipitation",
  days_back = 2
)
}
\arguments{
\item{data_path}{Path to the gridded rainall data}

\item{fun}{A function as character that the \code{exactext_extract}-function accepts}

\item{date}{A character that can be coerced directly to a date or an object of class \code{date}}

\item{nc_var}{The variable in the NetCDF you want to extract}

\item{days_back}{\code{integer vector} of days of antecedent rainfall you want to extract}

\item{sptial.obj}{An object of type \code{sf}. Either a point (or MULTIPOINT) or a polygon (or MULTIPOLYGON))}
}
\value{
a dataframe with either point extractions or columns of the aggregated values
}
\description{
Extract rainfall data
}
