% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ant_rainfall.R
\name{get_ant_rainfall}
\alias{get_ant_rainfall}
\title{Extract antecedent rainfall variables}
\usage{
get_ant_rainfall(
  data = NULL,
  id = "eidee",
  date_landslide = "date",
  vars = c(0, 1, 3, 5),
  dates_for_vars = NULL,
  parallel = T,
  rainfall_path = "/mnt/CEPH_PROJECTS/Proslide/PREC_GRIDS_updated/"
)
}
\arguments{
\item{date_landslide}{The name of the column with date information about the landslides.
Cells can either be NA or of class Date}
}
\description{
Extract antecedent rainfall variables
}
