% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cumulatitve_rainfall.R
\name{get_cumulative_rainfall}
\alias{get_cumulative_rainfall}
\title{Reshape the dataframe and create cumulative count column}
\usage{
get_cumulative_rainfall(res, cumsum = T)
}
\arguments{
\item{res}{an object as retured from `get_rainfall`}
}
\value{
a list of `dataframes` for each date with the cumulative amount for each day as column and the dates as another column.
}
\description{
Reshape the dataframe and create cumulative count column
}
