% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dates_ncdf}
\alias{get_dates_ncdf}
\title{Some helper functions}
\usage{
get_dates_ncdf(ncin, return_date_object = TRUE)
}
\arguments{
\item{ncin}{An object of class \code{ncdf4}}

\item{return_date_object}{\code{logical} If true returns a vector of
\code{date}-object, else vector of characterstring}
}
\description{
Some helper functions
}
