% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_annual_rainfall.R
\name{get_mean_annual_rainfall}
\alias{get_mean_annual_rainfall}
\title{Mean Annual precipitation per point}
\usage{
get_mean_annual_rainfall(
  data = NULL,
  map_path = "/mnt/CEPH_PROJECTS/Proslide/mean_annual_precipitation/map.tif",
  fun = c("mean"),
  crs = 32632
)
}
\arguments{
\item{data}{Object of class \code{sf}. Eihter POINT or POLYGON}

\item{map_path}{Path to raster of mean annual rainfall}

\item{fun}{Function to aggregated in case of POLYGON}
}
\description{
Get the mean annual rainfall for a spatial dataframe
}
