% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_events.R
\name{get_n_events}
\alias{get_n_events}
\title{Extract n rainfall events}
\usage{
get_n_events(
  point.data = NULL,
  n_dry = NULL,
  daily_threshold = NULL,
  n.events = NULL,
  ncores = 2,
  date_col = "date",
  id_col = "PIFF_ID",
  save = F,
  savePath = "/dev/null",
  path_rainfall = "/mnt/CEPH_PROJECTS/Proslide/PREC_GRIDS_updated/"
)
}
\arguments{
\item{point.data}{\code{sf} object of landsldies}

\item{n_dry}{Number of allowed dry days}

\item{daily_threshold}{Daily threshold that needs to be exceeded to be considered a rainy day}

\item{n.events}{how many rainfall events to extract for each landslide}

\item{ncores}{Number of cores working in parallel}

\item{date_col}{Character of the column that contains the dates of the landslides}

\item{id_col}{Character uniquely identifying each landslide}

\item{path_rainfall}{Character of where the rainfall data is saved}
}
\description{
This is a very inefficient way to extract a  number of rainfall events for each landslide.
It is inefficient because it checkes for each month if the number of rainfall events already exceeds the
specified number. If not, it will start from the beginning again...
Some landslides will have more events than specified, because landslides from the same dates are grouped
in order to only load the rasters once into memory. And extraction of rainfall values is done for each landslide
on that day until the last landslide exceeeds the specified number of rainfal events
}
