% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rainfall.R
\name{get_rainfall}
\alias{get_rainfall}
\title{Extract rainfall data}
\usage{
get_rainfall(
  data_path = "\\\\\\\\projectdata.eurac.edu/projects/Proslide/PREC_GRIDS/",
  spatial.obj = NULL,
  fun = NULL,
  dts = NULL,
  nc_var = "precipitation",
  seqq = TRUE,
  days_back = 2
)
}
\arguments{
\item{data_path}{Path to the gridded rainall data}

\item{fun}{A function to aggregate the data in case a polygon is passed as \code{spatial.obj}}

\item{dts}{Either a single object of type \code{Date} or a vector of at least two dates}

\item{seqq}{\code{logical} to create a sequence of days between two dates passed in \code{dts}}

\item{days_back}{\code{integer vector} of days of antecedent raindall you want to extract}

\item{sptial.obj}{An object of type \code{sf}. Either a point (or MULTIPOINT) or a polygon (or MULTIPOLYGON))}
}
\description{
Extract rainfall data
}
