% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rainfall_point_data.R
\name{get_rainfall_point_data}
\alias{get_rainfall_point_data}
\title{Get the rainfall-(events) for point data}
\usage{
get_rainfall_point_data(
  point.data = NULL,
  days_back = 10,
  daily_thresh = 0,
  n_dry = 0,
  parallel = TRUE,
  ncores = 8,
  last_event = FALSE,
  save = F,
  base_path = NULL,
  path_rainfall = "/mnt/CEPH_PROJECTS/Proslide/PREC_GRIDS_updated/"
)
}
\description{
Wrapper-funtion to calculate the rainfall, the cumulated sum, and
the events for point-data for specific dates (e.g. dates of landslides)
}
