% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yearly_rainfall.R
\name{get_yearly_rainfall}
\alias{get_yearly_rainfall}
\title{Get the summed yearly rainfall}
\usage{
get_yearly_rainfall(
  path_ncdf = "/mnt/CEPH_PROJECTS/Proslide/PREC_GRIDS_updated/",
  polygon = NULL,
  years = 1980:2018,
  by = "year",
  fun = NULL,
  aggre_fun = "mean"
)
}
\arguments{
\item{fun}{How to combine the individual values. E.G. the mean for each pixel for the years 1980, 1981 and 1982 requires the
the fun to be "mean". The maximum value for each month in the years 80,81,82 requires the \code{by}-parameter to be \code{month}
and the \code{fun}-parameter to be \code{max}}

\item{aggre_fun}{If you pass a polygon, the values need to be aggregated somehow}
}
\description{
This function takes either a polygon, or nothing (than it's pixel based) and calculates the sum of rainfall for each polygon
(each pixel) for each year
}
