% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iffi10_same_day.R
\name{iffi10_same_day}
\alias{iffi10_same_day}
\title{What slides happened on the same day?}
\usage{
iffi10_same_day(iffi10)
}
\arguments{
\item{iffi10}{A subset of the iffi10-database with already computed day-column}
}
\value{
A list where each element is a dataframe with all the slides from
that day
}
\description{
This function will group all of the points that happened on the
same day into a \code{data.frame} and put this dataframe into a list. Each
element of the list thus can be considered to be a day.
}
