% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aggregated_table.R
\name{make_aggregated_table}
\alias{make_aggregated_table}
\title{Calulate the the aggregated values for each polygon}
\usage{
make_aggregated_table(
  paths,
  polygon,
  years,
  by,
  daily_fun,
  monthly_fun,
  return.list = FALSE
)
}
\arguments{
\item{by}{either "year" or "month". If "year", the raster with the values for one year will be calculated and than the spatial
aggregation (with the function \code{aggre_fun}) will be applied. If "month", the spatial aggregation will be applied after each month}
}
\description{
This function will return a list of length n_years. With Either 12 (by = \code{month}) or 1 (by = \code{year}) elements per element
}
