% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cumsum.R
\name{make_cumsum}
\alias{make_cumsum}
\title{Will calculate the cumulative sum of the measurements over time}
\usage{
make_cumsum(df_wide, fun, days_back)
}
\description{
Will calculate the cumulative sum of the measurements over time
}
