% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cumulative_rainfall.R
\name{make_cumulative_rainfall}
\alias{make_cumulative_rainfall}
\title{Reshape the dataframe and create cumulative count column}
\usage{
make_cumulative_rainfall(res, cumsum = T)
}
\arguments{
\item{res}{an object as retured from `get_rainfall`}
}
\value{
a list of `dataframes` for each date with the cumulative amount for each day as as **one column** (long format)
}
\description{
Reshape the dataframe and create cumulative count column
}
