% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rasters.R
\name{make_rasters}
\alias{make_rasters}
\title{Calulate the aggregated Rasters}
\usage{
make_rasters(paths, years, by, daily_fun, monthly_fun, return.list = FALSE)
}
\description{
This function will return a list of length n_years. With Either 12 (by = \code{month}) or 1 (by = \code{year}) elements per element
}
