% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_dates_in_polygon.R
\name{slide_dates_in_polygon}
\alias{slide_dates_in_polygon}
\title{Extracts the days of the slides for each polygon}
\usage{
slide_dates_in_polygon(
  poly = NULL,
 
    iffi10 = "\\\\\\\\projectdata.eurac.edu/projects/Proslide/Landslides/Iffi_db_xxxx_to_2018/exportperEurac2020/Shapefiles/IFFI10_1.shp",
  epsg = 32632
)
}
\arguments{
\item{poly}{A path to a polygon spatial object}

\item{iffi10}{Path to the iffi10 shapefile with already precomputed columns in english and the date information}
}
\description{
Returns a \code{sf}-dataframe. For each polygon it will count the number of slides per date. This mean
that a polygon (its column \code{poly_id}) can potentially appear multiple times if on mulitple days slides happended in
this polygon. The computed column \code{slides_per_poly_date} counts the number of slides per date in that polygon.
This can then be fed into the function \code{get_rainfall_for_polygon} which is more or less a wrapper around the
function \code{ex_rainfall}
}
