% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rainfall_point_data.R
\name{get_rainfall_point_data}
\alias{get_rainfall_point_data}
\title{Get the rainfall-(events) for point data}
\usage{
get_rainfall_point_data(
  point.data = NULL,
  days_back = 10,
  daily_thresh = 0,
  n_dry = 0,
  parallel = TRUE,
  ncores = 8,
  last_event = FALSE,
  nle = FALSE,
  save = F,
  force = FALSE,
  base_path = NULL,
  id_landslide = NULL,
  antecedent = FALSE,
  path_rainfall = "/mnt/CEPH_PROJECTS/Proslide/PREC_GRIDS_updated/"
)
}
\arguments{
\item{last_event}{if True: Returns only the last event before the date of the landslide}

\item{nle}{if True: Returns all events. This will create a new column called \code{class} that either
stores \code{trigger} or \code{notrigger} depending on if there is any landslide correspondending in time
and space with a rainfall-event. if NLE == TRUE, all dates without any event will be omitted}

\item{id_landslide}{The name of a column that uniquely identifies each landslide}
}
\description{
Wrapper-funtion to calculate the rainfall, the cumulated sum, and
the events for point-data for specific dates (e.g. dates of landslides)
}
