% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aggregated_table.R
\name{make_aggregated_table}
\alias{make_aggregated_table}
\title{Calulate the the aggregated values for each polygon}
\usage{
make_aggregated_table(
  paths,
  polygon,
  years,
  by,
  daily_fun,
  monthly_fun,
  return.list = FALSE
)
}
\description{
This function will return a list of length n_years. With Either 12 (by = \code{month}) or 1 (by = \code{year}) elements per element
}
