% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_daily_rainfall_event.R
\name{reconstruct_daily_rainfall_events}
\alias{reconstruct_daily_rainfall_events}
\title{Extract rainfall Events per slide}
\usage{
reconstruct_daily_rainfall_events(d, n = 1, daily_thresh = 0.2, quiet = TRUE)
}
\arguments{
\item{d}{A dataframe with the extracted rainfall for one point or polygon. This dataframe needs to have a column
called \code{precip}}

\item{n}{The interval (in days) a dry period can still be part of a rainfall event}

\item{daily_thresh}{the minimum daily rainfall to be considered a day of rain}

\item{quiet}{Show print messages or not}
}
\value{
The same dataframe as inputted, just with the column \code{event} appended that shows the number of
rainfall event before the slide happened
}
\description{
Extracts the rainfall events that happened before a landslide, based on a predefined length ,
\code{n}, that can be dry days and still be part of a rainfall event.  Input usually is the result of the
function \code{ex_rainfall} for one slide with a unique `PIFF_ID`.
}
