% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_dates_in_polygon.R
\name{slide_dates_in_polygon}
\alias{slide_dates_in_polygon}
\title{Extracts the days of the slides for each polygon}
\usage{
slide_dates_in_polygon(poly = NULL, landsld = NULL, epsg = 32632)
}
\arguments{
\item{poly}{Path to the polygon shapefile}

\item{landsld}{Path to the landsld shapefile or object of class \code{sf}. Needs to have at least a column called \code{date_info}
(as in the classic iffi-data from the \code{iffitoR}-package) or column called \code{date} with dates of class \code{date}. If both do not
exist, the first column with the class \code{date} will be used to subset only the landslides with day-precise information}
}
\value{
An object of class \code{sf} which is (probably) of geometry-type polygon with the additional columns.
}
\description{
For each Polygon this function selects only the polygons that experienced a landsld. It also reports (in the column \code{slide_per_poly_date})
the number of slides on that date in that polygon.shapefile
}
