% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_dates_in_polygon.R
\name{slide_dates_in_polygon}
\alias{slide_dates_in_polygon}
\title{Extracts the days of the slides for each polygon}
\usage{
slide_dates_in_polygon(
  poly = NULL,
 
    iffi10 = "\\\\\\\\projectdata.eurac.edu/projects/Proslide/Landslides/Iffi_db_xxxx_to_2018/exportperEurac2020/Shapefiles/IFFI10_1.shp",
  epsg = 32632
)
}
\arguments{
\item{poly}{Path to the polygon shapefile}

\item{iffi10}{Path to the iffi10 shapefile with already precomputed columns in english}
}
\description{
Extracts the days of the slides for each polygon
}
